feat: Migrate project to use docker
This commit introduces Docker to the project, improving development and deployment processes.

The Makefile now leverages docker-compose for running different services, including the admin API, library, frontend, Postgres, and RabbitMQ. Project settings are adjusted to accommodate Docker, and necessary dependencies for Postgres and environment management are added.  The frontend Dockerfile is also updated to utilize the correct requirements file.